Remove requiredSources redirection
